Add FIPS code to counties in database
