+++ frontend/src/App.css
@@ -6,3 +6,15 @@
 .grid > div {
   padding: 6%;
 }
+
+.info {
+  position: absolute;
+  /* top: 10px;  */
+  right: 10px; 
+  background-color: rgba(255, 255, 255, 0.7); 
+  padding: 5px;
+  border-radius: 5px;
+  font-size: 14px;
+  width: 150px;
+  z-index: 1000
+}
