+++ frontend/src/components/index.ts
@@ -2,6 +2,7 @@ export {default as Navbar } from "./Navbar";
 export {default as Dropdown} from "./Dropdown";
 export {default as DropdownList} from "./DropdownList";
 export {default as SkillSearchMap} from "./SkillSearchMap";
+export {default as Choropleth} from "./Choropleth";
 export {default as DashboardMap} from "./DashboardMap";
 export {default as SkillChart} from "./SkillChart";
 export {default as JobList} from "./JobList";
